tracing for Rusoto ops (#15)
* proof-of-concept tracing for Rusoto ops

This includes the `rk=` keys being operated on. My assumption is they're
never sensitive. If that's not true, maybe we need a trait that callers
can supply to show the appropriate redacted form.

It also includes errors and consumed RCU/WCU.

Currently implemented only for get batches and transactions:

* I decided to create a separate span for each operation within the
  batch/transaction, lacking any guidance from e.g. the opentelemetry
  spec: https://github.com/open-telemetry/opentelemetry-specification/issues/1548

* The non-batch operations are easy, just tedious. I'll do them if folks
  like the idea of tracing this.

* use a new trait for redactable hash keys

* commit lockfile

This addresses CI bitrot such as the following:

```
error: package `time v0.3.22` cannot be built because it requires rustc 1.65.0 or newer, while the currently active rustc version is 1.62.1
```

Cargo official docs currently recommend against committing a lockfile
for libraries, but I think they're wrong, and they're considering
changing this recommendation:
https://rust-lang.zulipchat.com/#narrow/stream/246057-t-cargo/topic/Commit.20lockfiles.20for.20libs.3F/near/323754474

Applications don't use the library's `Cargo.lock`, but it allows
stability for CI and documents known-working versions.

* assume static strings are non-sensitive

* improve ExplicitKey usability

(after prototyping the av-service commit)

* bulk out the tracing impl a bit

* trace the single-op variants
* have those all delegate to _impl methods instead of to others, so that
  the `#[tracing::instrument]` annotations do the right things
* use `#[tracing::instrument(err(Display))]` rather than manually
  set the error annotation

* use Rust 1.70

(av-service and friends already are on this version)