fix rcu and wcu reporting
This was working just for the `exec_atomic_write` case. DynamoDB
separately reports rcu and wcu there. Apparently for other operations
it doesn't do that, e.g.

```
aws-vault exec fusion-dev-ro -- aws dynamodb batch-get-item \
    --return-consumed-capacity TOTAL \
    --request-items '{"av-service-dev-DynamoDBTable-1651M3QJDASG3": {"Keys": [

{
        "hk": {"B": "YXZfc3RyZWFtX2FjdGl2aXR5X2J5X2V2ZW50X2lkOna4ZK/B1ayXTkA8usO5Ret8uh4w"},
        "rk": {"B": "Xw==" }}

]}}'
...
    "ConsumedCapacity": [
        {
            "TableName": "av-service-dev-DynamoDBTable-1651M3QJDASG3",
            "CapacityUnits": 1.0
        }
    ]
```

so look at the single `CapacityUnits` number and report it as rcu or
wcu depending on the operation we're performing.